third_party/blink/web_tests/platform/generic/fast/dom/TreeWalker/filter-throw-expected.txt

Test JS objects as NodeFilters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing with object filter
Test with filter function
PASS walker.firstChild(); threw exception filter exception.
PASS walker.currentNode.id is 'root'
PASS walker.nextNode();walker.currentNode.id threw exception filter exception.
PASS walker.currentNode.id is 'root'

Test with filter object
PASS walker.firstChild(); threw exception filter exception.
PASS walker.currentNode.id is 'root'
PASS walker.nextNode();walker.currentNode.id threw exception filter exception.
PASS walker.currentNode.id is 'root'
PASS successfullyParsed is true

TEST COMPLETE
